Always use Services.wm for nsIWindowMediator
No need to assign it to `Zotero.ZotFile.wm`, and after f1ef7cbbef6 an
error was being thrown if you right-clicked on an item before `init()`
finished waiting for `schemaUpdatePromise`.

There are probably other things in `init()` that need to finish before
various operations will work, so the context-menu options should
probably be disabled until `schemaUpdatePromise` is done.