Fix various errors (#431)
* Fix attachment rename when attachment has a tag

`addTags()` was removed in Zotero 5

* Fix "this is undefined" error in ospath_win.jsm

`OS.Path.normalize()` uses `this`, so passing `OS.Path.normalize`
as a callback without rebinding will fail. This uses an arrow function
instead.

I believe this would've caused `renameAttachment()` to fail when going
from linked to imported for multiple files (and I think there've been
reports of that). In other cases it would've just been logged, since the
async parent function, `removeEmptyFolders()`, wasn't waited for. I've
now changed that to a `yield` so errors are caught properly.

* Fix "attachmentSyncedHash can only be set for stored files" logged error

This was logged when moving from one linked location to another.
`attachmentSyncedHash` and `attachmentSyncState` can't be set on linked
attachments.

This also stops updating the file timestamp, which I don't think is
necessary. We're only doing that in `renameAttachmentFile()` to trigger
Zotero file syncing.