Zotero 6 compatibility (#573)
* Zotero 6 compatibility: Don't lose annotations and other attachment data

When converting a linked-file attachment to a stored-file attachment or
vice versa, transfer the child annotations and other attachment data
(relations, full-text-index content) to the new item rather the
discarding it.

* Zotero 6 compatibility: Don't show Extract Annotations functionality

This hides the Extract Annotations menu option and the annotation
extraction settings in the Advanced preferences when running Zotero 6.

Zotero 6 provides various options on its own [1] to create notes from
both embedded PDF annotations and the annotations created in Zotero's
new PDF reader. ZotFile doesn't know anything about Zotero annotations,
which has been causing a lot of user confusion during the beta, and that
would only increase once Zotero 6 is released and the built-in PDF
reader becomes the default option.

Zotero's annotation extraction is also much more flexible, since it
includes much more internal metadata in the notes: colors and citations
can be toggled on and off in the notes and will be customizable in
exports, rich UI can be displayed for various actions (opening the PDF
and highlighting the annotation, jumping back to the library item,
restoring a deleted citation), it will be possible to automatically
replace citations with pandoc-style citations for Markdown export, the
citation format displayed in notes could be made customizable, etc.

Any remaining annotation-related functionality that Zotero doesn't
currently support (e.g., `.pdfExtraction.NoteHtmlTagStart` and similar)
should either be updated in ZotFile to modify Zotero's functionality or
proposed/contributed as patches to Zotero itself.

This patch doesn't remove the option to automatically extract
annotations when retrieving a tablet file, though that would ideally be
updated to just trigger Zotero's built-in annotation extraction so that
the above benefits would apply and so that all of ZotFile's extraction
code could be removed. Zotero displays PDF bookmarks in the reader
sidebar, but if we think that's a feature people use we could probably
add ToC/bookmark extraction to Zotero proper as well.

[1] https://www.zotero.org/support/pdf_reader#adding_annotations_to_notes

* Zotero 6 compatibility: Update maxVersion in install.rdf