Fix hang during rename-and-move in Zotero 6
Resulting in not all annotations being transferred and the old
attachment not being deleted

Regression after zotero/zotero@734057ff9b, which added a transaction
requirement to `Zotero.Items.moveChildItems()` but tried to preserve
backwards compatibility with ZotFile. Unfortunately it had a bug, and a
race condition with another transacation during the process results in a
hang anyway if `moveChildItems()` isn't called in a transaction, so just
change ZotFile to follow the new requirement.

This also fixes `saveTx()` being called without an await.

Fixes #593